init: mongodb 모듈 설정 및 임시 컨트롤러 설정
