+++ Backend/api/src/main/java/com/example/ApiApplication.java
@@ -3,12 +3,7 @@ package com.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
-//@EnableJpaRepositories({"com.example.core.repository"})
-//@EntityScan({"com.example.core.entity"})
 @SpringBootApplication
 public class ApiApplication {
     public static void main(String[] args) {
