+++ Backend/api/src/main/java/com/example/service/ProductService.java
@@ -1,8 +1,8 @@
 package com.example.service;
 
-import com.example.domain.product.Product;
-import com.example.repository.product.ProductReader;
-import com.example.repository.product.ProductWriter;
+import com.example.mysql.domain.product.Product;
+import com.example.mysql.repository.product.ProductReader;
+import com.example.mysql.repository.product.ProductWriter;
 import com.example.request.product.CreateProductRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,7 +17,7 @@ public class ProductService {
     private final ProductReader productReader;
 
     @Transactional(value = "productTransactionManager")
-    public Product createProduct(CreateProductRequest request){
+    public Product createProduct(CreateProductRequest request) {
         Product product = Product.builder()
                 .productName(request.getProductName())
                 .build();
@@ -25,7 +25,7 @@ public class ProductService {
     }
 
     @Transactional(value = "productTransactionManager")
-    public Product findByProductId(Long productId){
+    public Product findByProductId(Long productId) {
         return productReader.findByProductId(productId);
     }
 }
