--- Backend/persistence/mongodb/src/main/java/com/example/config/CartDBConfig.java
@@ -1,4 +1,0 @@
-package com.example.config;
-
-public class CartDBConfig {
-}
