+++ Backend/persistence/mongodb/src/main/java/com/example/mongodb/repository/CartRepository.java
@@ -1,0 +1,9 @@
+package com.example.mongodb.repository;
+
+import com.example.mongodb.entity.Cart;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CartRepository extends MongoRepository<Cart, String> {
+}
