--- Backend/persistence/mysql/src/main/java/com/example/config/EntityManagerFactoryBuilderConfig.java
@@ -1,19 +1,0 @@
-package com.example.config;
-
-import org.springframework.boot.orm.jpa.EntityManagerFactoryBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.orm.jpa.vendor.Database;
-import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
-
-import java.util.HashMap;
-
-@Configuration
-public class EntityManagerFactoryBuilderConfig {
-    @Bean
-    public EntityManagerFactoryBuilder entityManagerFactoryBuilder() {
-        HibernateJpaVendorAdapter vendorAdapter = new HibernateJpaVendorAdapter();
-        vendorAdapter.setDatabase(Database.MYSQL);
-        return new EntityManagerFactoryBuilder(vendorAdapter, new HashMap<>(), null);
-    }
-}
