--- Backend/persistence/mysql/src/main/java/com/example/config/RatingDBConfig.java
@@ -1,4 +1,0 @@
-package com.example.config;
-
-public class RatingDBConfig {
-}
