+++ Backend/persistence/mysql/src/main/java/com/example/mysql/config/RatingDBConfig.java
@@ -1,0 +1,4 @@
+package com.example.mysql.config;
+
+public class RatingDBConfig {
+}
