+++ Backend/persistence/mysql/src/main/java/com/example/mysql/domain/product/Product.java
@@ -1,0 +1,15 @@
+package com.example.mysql.domain.product;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@Builder
+@AllArgsConstructor
+public class Product {
+    private Long productId;
+    private String productName;
+}
