+++ Backend/persistence/mysql/src/main/java/com/example/mysql/domain/user/User.java
@@ -1,0 +1,15 @@
+package com.example.mysql.domain.user;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@Builder
+@AllArgsConstructor
+public class User {
+    private Long userId;
+    private String userName;
+}
