+++ Backend/persistence/mysql/src/main/java/com/example/mysql/entity/product/ProductEntity.java
@@ -1,0 +1,41 @@
+package com.example.mysql.entity.product;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import java.time.Instant;
+
+@Table(name = "products")
+@Getter
+@Entity
+@NoArgsConstructor
+@Builder
+@AllArgsConstructor
+public class ProductEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "product_id")
+    private Long productId;
+
+    @Column(name = "product_name")
+    private String productName;
+
+    @Column(name = "created_dt")
+    private Instant createdDt;
+    @Column(name = "updated_dt")
+    private Instant updatedDt;
+
+    @PrePersist
+    public void prePersist() {
+        this.createdDt = Instant.now();
+        this.updatedDt = Instant.now();
+    }
+
+    @PreUpdate
+    public void preUpdate() {
+        this.updatedDt = Instant.now();
+    }
+}
