+++ Backend/persistence/mysql/src/main/java/com/example/mysql/repository/product/ProductEntityRepository.java
@@ -1,0 +1,7 @@
+package com.example.mysql.repository.product;
+
+import com.example.mysql.entity.product.ProductEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductEntityRepository extends JpaRepository<ProductEntity, Long> {
+}
