+++ Backend/persistence/mysql/src/main/java/com/example/mysql/repository/product/ProductReader.java
@@ -1,0 +1,7 @@
+package com.example.mysql.repository.product;
+
+import com.example.mysql.domain.product.Product;
+
+public interface ProductReader {
+    Product findByProductId(Long productId);
+}
