+++ Backend/persistence/mysql/src/main/java/com/example/mysql/repository/product/ProductReaderImpl.java
@@ -1,0 +1,21 @@
+package com.example.mysql.repository.product;
+
+import com.example.mysql.domain.product.Product;
+import com.example.mysql.entity.product.ProductEntity;
+import com.example.mysql.entity.product.ProductMapper;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class ProductReaderImpl implements ProductReader {
+    private final ProductEntityRepository productEntityRepository;
+    private final ProductMapper productMapper;
+
+    @Override
+    public Product findByProductId(Long productId) {
+        ProductEntity productEntity = productEntityRepository.findById(productId)
+                .orElseThrow(IllegalArgumentException::new);
+        return productMapper.toDomain(productEntity);
+    }
+}
