+++ Backend/persistence/mysql/src/main/java/com/example/mysql/repository/product/ProductWriterImpl.java
@@ -1,0 +1,19 @@
+package com.example.mysql.repository.product;
+
+import com.example.mysql.domain.product.Product;
+import com.example.mysql.entity.product.ProductMapper;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class ProductWriterImpl implements ProductWriter {
+    private final ProductEntityRepository productEntityRepository;
+    private final ProductMapper productMapper;
+
+    public Product create(Product product) {
+        return productMapper.toDomain(productEntityRepository.save(productMapper.toEntity(product)));
+    }
+
+
+}
