+++ Backend/persistence/mysql/src/main/java/com/example/mysql/repository/user/UserEntityRepository.java
@@ -1,0 +1,7 @@
+package com.example.mysql.repository.user;
+
+import com.example.mysql.entity.user.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserEntityRepository extends JpaRepository<UserEntity, Long> {
+}
