+++ Backend/persistence/mysql/src/main/java/com/example/mysql/repository/user/UserReaderImpl.java
@@ -1,0 +1,10 @@
+package com.example.mysql.repository.user;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class UserReaderImpl implements UserReader {
+    private UserEntityRepository userEntityRepository;
+}
