+++ Backend/persistence/mysql/src/main/java/com/example/mysql/repository/user/UserWriter.java
@@ -1,0 +1,4 @@
+package com.example.mysql.repository.user;
+
+public interface UserWriter {
+}
