+++ Backend/persistence/mysql/src/main/java/com/example/mysql/repository/user/UserWriterImpl.java
@@ -1,0 +1,11 @@
+package com.example.mysql.repository.user;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class UserWriterImpl implements UserWriter {
+    private UserEntityRepository userEntityRepository;
+
+}
