--- Backend/persistence/mysql/src/main/java/com/example/repository/product/ProductEntityRepository.java
@@ -1,7 +1,0 @@
-package com.example.repository.product;
-
-import com.example.entity.product.ProductEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductEntityRepository extends JpaRepository<ProductEntity, Long> {
-}
