--- Backend/persistence/mysql/src/main/java/com/example/repository/product/ProductWriter.java
@@ -1,7 +1,0 @@
-package com.example.repository.product;
-
-import com.example.domain.product.Product;
-
-public interface ProductWriter {
-    Product create(Product product);
-}
