--- Backend/persistence/mysql/src/main/java/com/example/repository/product/ProductWriterImpl.java
@@ -1,19 +1,0 @@
-package com.example.repository.product;
-
-import com.example.domain.product.Product;
-import com.example.entity.product.ProductMapper;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class ProductWriterImpl implements ProductWriter{
-    private final ProductEntityRepository productEntityRepository;
-    private final ProductMapper productMapper;
-
-    public Product create(Product product){
-        return productMapper.toDomain(productEntityRepository.save(productMapper.toEntity(product)));
-    }
-
-
-}
