--- Backend/persistence/mysql/src/main/java/com/example/repository/user/UserEntityRepository.java
@@ -1,7 +1,0 @@
-package com.example.repository.user;
-
-import com.example.entity.user.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserEntityRepository extends JpaRepository<UserEntity, Long> {
-}
