--- Backend/persistence/mysql/src/main/java/com/example/repository/user/UserReaderImpl.java
@@ -1,10 +1,0 @@
-package com.example.repository.user;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class UserReaderImpl implements UserReader{
-    private UserEntityRepository userEntityRepository;
-}
