--- Backend/persistence/mysql/src/main/java/com/example/repository/user/UserWriter.java
@@ -1,4 +1,0 @@
-package com.example.repository.user;
-
-public interface UserWriter {
-}
