--- Backend/persistence/mysql/src/main/java/com/example/repository/user/UserWriterImpl.java
@@ -1,11 +1,0 @@
-package com.example.repository.user;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class UserWriterImpl implements UserWriter{
-    private UserEntityRepository userEntityRepository;
-
-}
