--- Backend/persistence/mysql/src/main/resources/application.yml
@@ -1,32 +1,0 @@
-server:
-  port: 8081
-
-spring:
-  datasource:
-    # mysql
-    user:
-      driver-class-name: com.mysql.cj.jdbc.Driver
-      url: jdbc:mysql://localhost:13307/userDB  #  url: jdbc:mysql://user-mysql:3306/userDB
-      username: root
-      password: 1234
-    product:
-      driver-class-name: com.mysql.cj.jdbc.Driver
-      url: jdbc:mysql://localhost:13306/productDB # url: jdbc:mysql://product-mysql:3306/productDB
-      username: root
-      password: 1234
-
-
-  jpa:
-    hibernate:
-      ddl-auto: none
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.MySQL8Dialect
-        show_sql: true
-    open-in-view: false
-
-logging:
-  level:
-    org.hibernate.sql: debug
-    org.hibernate.type: trace
-
